siteController moved to single file
