+++ rct-demo/app/public/utils/handleClick.js
